Bugfix, bot wouldn't leave voiceChannel :zap:

+++ commands/play.js
@@ -38,6 +38,8 @@ module.exports = {
             return;
         }
 
+        if(voiceChannel == undefined || voiceChannel == null) return;
+
         message.channel.bulkDelete(1);
 
         message.reply("we here at Georg Music will now play your requested audio at volume " + vol +   ": ");
